016 - fix: replace interface
