+++ src/app/core/description-service.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Internship } from '../types';
+import {Internship, Training} from '../types';
 
 @Injectable({
   providedIn: 'root'
@@ -8,8 +8,33 @@ import { Internship } from '../types';
 
 export class DescriptionService {
   basePath = 'http://localhost:8080/api/internships';
-  constructor(private http: HttpClient) { 
-
+  trainings: Training[] = [
+    {
+      country: 'Belarus',
+      city: 'Minsk',
+      form: 'Online',
+      technology: 'HTML',
+      img: 'assets/icons/html.svg',
+      date: '25 april 2021',
+    },
+    {
+      country: 'Ukraine',
+      city: 'Kiev',
+      form: 'Online',
+      technology: 'CSS',
+      img: 'assets/icons/css.svg',
+      date: '15 april 2021',
+    },
+    {
+      country: 'Russia',
+      city: 'Moscow',
+      form: 'Online',
+      technology: 'JavaScript',
+      img: 'assets/icons/JS.svg',
+      date: '3 april 2021',
+    }
+  ];
+  constructor(private http: HttpClient) {
   }
 
   getInternshipList(){
@@ -18,5 +43,9 @@ export class DescriptionService {
 
   getInternshipById(id: string){
     return this.http.get<Internship>(`${this.basePath}/${id}`);
+  }
+
+  getTrainingsLocal(){
+    return this.trainings;
   }
 }
