+++ src/app/homepage/training/training.component.html
@@ -3,6 +3,6 @@
   <img [src]="training.img" alt="" class="training_course_img"/>
   <div class="training_course_technology">Technology: {{training.technology}}</div>
   <div class="training_course_name">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quisquam, quos?</div>
-  <div class="training_course_dateRegistration">{{training.data}}</div>
+  <div class="training_course_dateRegistration">{{training.date}}</div>
   <button mat-raised-button color="primary" class="btn-training">show more</button>
 </div>
