+++ src/app/types/internship.ts
@@ -1,22 +1,32 @@
 export interface Internship {
-    capacity: number;
-    countryList: Location[];
-    description: string;
-    endDate: string;
-    id: string;
-    internshipFormat: string;
-    name: string;
-    publicationDate: string;
-    registrationEndDate: string;
-    registrationStartDate: string;
-    requirements: string;
-    startDate: string;
-    techSkills: string;
-    technologyList: string[];
+  capacity: number;
+  countryList: Location[];
+  description: string;
+  endDate: string;
+  id: string;
+  internshipFormat: string;
+  name: string;
+  publicationDate: string;
+  registrationEndDate: string;
+  registrationStartDate: string;
+  requirements: string;
+  startDate: string;
+  techSkills: string;
+  technologyList: string[];
 }
 
-  export interface Location {
-      city: string;
-      country: string;
-      id: string;
-  }
+export interface Location {
+  city: string;
+  country: string;
+  id: string;
+}
+
+export interface Training {
+  id?: string;
+  country: string;
+  city: string;
+  form: string;
+  technology: string;
+  img: string;
+  date?: string;
+}
